Ingest scripts utilizing separate .py files for shared functions and scripts; start with dataset ingest first; first commit
