+++ ingest/Maybe.py
@@ -1,0 +1,55 @@
+__author__ = 'Stephan'
+
+from itertools import chain
+
+class Maybe:
+    def __init__(self, v=None):
+        self.value = v
+
+    @staticmethod
+    def nothing():
+        return Maybe()
+
+    @staticmethod
+    def of(t):
+        return Maybe(t)
+
+    def reduce(self, action):
+        return Maybe.of(functools.reduce(action, self.value)) if self.value else Maybe.nothing()
+
+    def stream(self):
+        return Maybe.of([self.value]) if self.value is not None else Maybe.nothing()
+
+    def map(self, action):
+        return Maybe.of(chain(map(action, self.value))) if self.value is not None else Maybe.nothing()
+
+    def flatmap(self, action):
+        return Maybe.of(chain.from_iterable(map(action, self.value))) if self.value is not None else Maybe.nothing()
+
+    def andThen(self, action):
+        return Maybe.of(action(self.value)) if self.value is not None else Maybe.nothing()
+
+    def orElse(self, action):
+        return Maybe.of(action()) if self.value is None else Maybe.of(self.value)
+
+    def do(self, action):
+        if self.value:
+            action(self.value)
+        return self
+
+    def filter(self, action):
+        return Maybe.of(filter(action, self.value)) if self.value is not None else Maybe.nothing()
+
+    def followedBy(self, action):
+        return self.andThen(lambda _: action)
+
+    def one(self):
+        try:
+            return Maybe.of(next(self.value)) if self.value is not None else Maybe.nothing()
+        except StopIteration:
+            return Maybe.nothing()
+        except TypeError:
+            return self
+
+    def list(self):
+        return list(self.value) if self.value is not None else []
